# Performance improvements: use `multiSearchAny` instead of `IN` to support Clickhouse projections for logs volumes