@SpencerTorres 2024-11-27
This looks good, I'll run it locally to verify it works for our other use cases

@andreygubarev  Can you add a changelog entry? Thanks!